+++ webpack/develop.js
@@ -33,6 +33,7 @@ const config = merge(common, {
         open: true,
         port: 33330,
         hot: true,
+        // host: '0.0.0.0',
         // hot 和 hotOnly 的区别是在某些模块不支持热更新的情况下，前者会自动刷新页面，后者不会刷新页面，而是在控制台输出热更新失败
         hotOnly:true,
         proxy: {
@@ -68,7 +69,7 @@ const config = merge(common, {
             test: /\.less$/,
             use: [
                 {
-                    loader: MiniCssExtractPlugin.loader,
+                    loader: 'style-loader',
                     options: {
                         hmr: true
                     },